% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest.R
\name{ftest.pow.norm}
\alias{ftest.pow.norm}
\title{Area under the \code{ftest} power function}
\usage{
ftest.pow.norm(tau, n.of.y, p, support = c(0, Inf))
}
\arguments{
\item{tau}{Upper boundary point of the equivalence region}

\item{p}{Number of variables}

\item{support}{Support of the truncated power function (vector of dimension 2).}
}
\description{
This function computes the area under the power function \code{ftest.pow}.
}
\seealso{
\code{\link{ftest.pow}}
}

