\name{ma.a2nu}
\alias{ma.a2nu}
\title{Transform \code{a} to the autocorrelation of the MA(1) process}
\usage{
ma.a2nu(a)
}
\arguments{
  \item{a}{\code{a} parameter}
}
\value{
Autocorrelation parameter of the MA(1) process
}
\description{
Transform \code{a} to the autocorrelation of the MA(1)
process
}

