\name{ma.a2rho}
\alias{ma.a2rho}
\title{Transform \code{a} to the autocorrelation error of the MA(1) process}
\usage{
ma.a2rho(a, ax = 0)
}
\arguments{
  \item{a}{\code{a} parameter}

  \item{ax}{MLE of the autocorrelation of the observed
  summary values}
}
\value{
Autocorrelation error parameter of the MA(1) process
}
\description{
Transform \code{a} to the autocorrelation error of the
MA(1) process
}

