\name{ma.nu2a}
\alias{ma.nu2a}
\title{Transform the autocorrelation of the MA(1) process into the \code{a} parameter}
\usage{
ma.nu2a(x)
}
\arguments{
  \item{x}{Autocorrelation parameter of the MA(1) process}
}
\value{
\code{a} parameter
}
\description{
Transform the autocorrelation of the MA(1) process into the
\code{a} parameter
}

