\name{ma.rho2a}
\alias{ma.rho2a}
\title{Transform the autocorrelation error of the MA(1) process to the \code{a} parameter}
\usage{
ma.rho2a(x, ax = 0)
}
\arguments{
  \item{x}{Autocorrelation error parameter}

  \item{ax}{MLE of the autocorrelation of the observed
  summary values}
}
\value{
\code{a} parameter
}
\description{
Transform the autocorrelation error of the MA(1) process to
the \code{a} parameter
}

