\name{ma.rho2sig2}
\alias{ma.rho2sig2}
\title{Transform the variance error of the MA(1) process to \code{sig2}}
\usage{
ma.rho2sig2(x, a, vx = 1)
}
\arguments{
  \item{x}{the variance error of the MA(1) process}

  \item{a}{\code{a} parameter}

  \item{vx}{MLE of the variance of the observed summary
  values}
}
\value{
\code{sig2} parameter
}
\description{
Transform the variance error of the MA(1) process to
\code{sig2}
}

