\name{ma.sig22rho}
\alias{ma.sig22rho}
\title{Transform \code{sig2} to the variance error of the MA(1) process}
\usage{
ma.sig22rho(s2, a, vx = 1)
}
\arguments{
  \item{s2}{\code{sig2} parameter}

  \item{a}{\code{a} parameter}

  \item{vx}{MLE of the variance of the observed summary
  values}
}
\value{
Variance error of the MA(1) process
}
\description{
Transform \code{sig2} to the variance error of the MA(1)
process
}

