% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutost.R
\name{mutost.pow.norm}
\alias{mutost.pow.norm}
\title{Area under the \code{mutost} power function}
\usage{
mutost.pow.norm(df, s.of.T, c.u, tau.u = NA, alpha = NA, norm = 1,
  support = c(-Inf, Inf), log = FALSE)
}
\arguments{
\item{df}{Degrees of freedom}

\item{s.of.T}{Standard deviation of the test statistic}

\item{c.u}{Upper boundary point of the critical region}

\item{tau.u}{Upper boundary point of the equivalence region}

\item{alpha}{Level of the equivalence test}

\item{norm}{Normalization constant for the truncated power function.}

\item{support}{Support of the truncated power function (vector of dimension 2).}

\item{log}{If \code{TRUE}, the power function is returned on the log scale.}
}
\description{
This function computes the area under the power function \code{mutost.pow}.
}
\references{
http://arxiv.org/abs/1305.4283
}
\seealso{
\code{\link{mutost.pow}}
}

