\docType{data}
\name{NABC.DEFAULT.ANS}
\alias{NABC.DEFAULT.ANS}
\title{this file contains all R functions of the abc-n package}
\format{Named num [1:13] 0 50 1 NA NA NA 0 0 0 1 ...
 - attr(*, "names")= chr [1:13] "lkl" "error" "pval" "link.mc.obs" ...}
\usage{
  NABC.DEFAULT.ANS
}
\description{
  this file contains all R functions of the abc-n package
}
\keyword{datasets}

