% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratetest.R
\name{ratetest.pow}
\alias{ratetest.pow}
\title{\code{ratetest} power function}
\usage{
ratetest.pow(rho, c.l, c.u, m, norm = 1, support = c(0, Inf), log = FALSE)
}
\arguments{
\item{rho}{Vector of error quantiles}

\item{c.l}{Lower boundary point of the critical region (equivalent to the lower ABC tolerance \code{epsilon^-})}

\item{c.u}{Upper boundary point of the critical region (equivalent to the upper ABC tolerance \code{epsilon^+})}

\item{m}{Number of simulated values}

\item{norm}{Normalization constant for the truncated power function}

\item{support}{Support of the truncated power function}

\item{log}{If \code{TRUE}, the power function is returned on the log scale.}
}
\description{
Computes the power function of the two-sided, one-sample gamma test for testing if simulated and observed 
summary values occur at similar rates. This test is applicable when the observed and simulated summary values are Exponentially
distributed, or when Exponentiality cannot be rejected. The testing problem is described in terms of the scale parameter of the Exponential distribution, the reciprocal of the
rate parameter.
}
\note{
The power function can be truncated to \code{support} and then standardized with \code{norm}.
If one of these is set, the other must be provided too.
}
\examples{
n.of.y <- 40

# compute ABC tolerances
cali	<- ratetest.calibrate(tau.l=1/2, tau.u=2, n.of.y=n.of.y, what='CR', alpha=0.01)
# compute the power for the range (0.1, 3)
rho  	<- seq(0.1, 3, len=1024)
tmp		<- data.frame(rho=rho, power=ratetest.pow(rho, cali['c.l'], cali['c.u'], m=n.of.y))

library(ggplot2)
p 		<- ggplot(tmp,aes(x=rho,y=power)) + geom_line() + labs(y='Power\\n(ABC acceptance probability)')
print(p)



}
\seealso{
\code{\link{ratetest.calibrate}}, \code{\link{ratetest.pow.norm}}
}

