% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsftest.R
\name{tsftest.pow}
\alias{tsftest.pow}
\title{\code{tsftest} power function}
\usage{
tsftest.pow(rho, cl, cu, n.of.y, p, support = c(0, Inf), log = FALSE,
  norm = 1)
}
\arguments{
\item{rho}{Vector of quantiles}

\item{cl}{Lower boundary point of the critical region}

\item{cu}{Upper boundary point of the critical region}

\item{n.of.y}{Number of replicate simulations}

\item{p}{Number of variables}

\item{support}{Support of the truncated power function (vector of dimension 2).}

\item{log}{If \code{TRUE}, the power function is returned on the log scale.}
}
\description{
Compute the power of the two-sided one-sample multivariate equivalence test for population means of multivariate normal summary values with unknown population variance.
}
\note{
The power function can be truncated to \code{support}.
}
\examples{
# power function of the two-sided F-test, to test equality of means for multivariate
# normal samples with unknown covariance matrix

n		<- 10
p		<- 3
cu		<- 8
rho 	<- seq(0, 2, length = 1024)
tmp <- lapply(c(0.01, 0.5, 1, 2, 3, 4, 5, 6, 7), function(cl)
		{
			data.table(cl = as.factor(cl), cu=cu, rho = rho, power = tsftest.pow(rho, cl, cu, n, p))
		})
tmp	<- do.call('rbind', tmp)
pp <- ggplot(tmp, aes(x = rho, y = power, colour = cl, group = cl)) + geom_line() + labs(y = 'Power\\n(ABC acceptance probability)')
print(pp)
}
\references{
http://arxiv.org/abs/1305.4283
}

