% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vartest.pow.norm}
\alias{vartest.pow.norm}
\title{Area under the \code{chi2} power function}
\usage{
vartest.pow.norm(scale, df, c.l, c.u, trafo = 1, support = c(0, Inf))
}
\arguments{
\item{scale}{Scaling parameter; typically set to the number of observed summary values}

\item{df}{Degrees of freedom; typically set to \code{m-1} where \code{m} is the number of simulated summary values}

\item{c.l}{Lower boundary point of the critical region (equivalent to the lower ABC tolerance \code{epsilon^-})}

\item{c.u}{Upper boundary point of the critical region (equivalent to the upper ABC tolerance \code{epsilon^+})}

\item{trafo}{Parameter transformation to translate the power}

\item{support}{Support of the truncated power function}
}
\description{
This function computes the area under the power function \code{vartest.pow.norm}.
}
\seealso{
\code{\link{vartest.pow}}
}

