% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztest.R
\name{ztest.pow.norm}
\alias{ztest.pow.norm}
\title{Area under the \code{ztest} power function}
\usage{
ztest.pow.norm(tau.u, sigma, alpha = 0.01, support = c(-Inf, Inf))
}
\arguments{
\item{tau.u}{Upper boundary point of the equivalence region}

\item{sigma}{Standard deviation of the test statistic.}

\item{alpha}{Level of the equivalence test}

\item{support}{Support of the truncated power function (vector of dimension 2).}
}
\description{
This function computes the area under the power function \code{ztest.pow}.
}
\seealso{
\code{\link{ztest.pow}}
}

